+++ src/Indice.Features.Messages.Core/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.4.4] - 2023-10-04
+### Added
+- ContactRetainPeriodInDays option to keep in sync a contact with the identity system. 
+  After the configured period of time the system patches and updates the contact with the latest values.
+
 ## [7.4.1] - 2023-09-22
 ### Changed
 - CampaignId is returned in PushNotification data in property "messageId".